[Edit] Migration

+++ db/migrate/20210918022530_create_post_item_genres.rb
@@ -1,8 +1,8 @@
 class CreatePostItemGenres < ActiveRecord::Migration[5.2]
   def change
     create_table :post_item_genres do |t|
-      t.integer :genre_id, null: false
-      t.integer :post_item_id, null: false
+      t.bigint :genre_id, null: false
+      t.bigint :post_item_id, null: false
       t.timestamps null: false
     end
       add_foreign_key :post_item_genres, :genres, column: :genre_id
