+++ db/migrate/20210929062353_create_favorites.rb
@@ -1,8 +1,8 @@
 class CreateFavorites < ActiveRecord::Migration[5.2]
   def change
     create_table :favorites do |t|
-      t.integer :user_id, null: false
-      t.integer :post_item_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :post_item_id, null: false
       t.timestamps
     end
       add_foreign_key :favorites, :users, column: :user_id
