+++ db/migrate/20211003022005_create_comments.rb
@@ -1,8 +1,8 @@
 class CreateComments < ActiveRecord::Migration[5.2]
   def change
     create_table :comments do |t|
-      t.integer :user_id, null: false
-      t.integer :post_item_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :post_item_id, null: false
       t.text :content, null: false
       t.timestamps null: false
     end
