+++ db/migrate/20211005133704_create_entries.rb
@@ -1,8 +1,8 @@
 class CreateEntries < ActiveRecord::Migration[5.2]
   def change
     create_table :entries do |t|
-      t.integer "user_id", null: false
-      t.integer "chat_room_id", null: false
+      t.bigint "user_id", null: false
+      t.bigint "chat_room_id", null: false
       t.timestamps
     end
       add_foreign_key :entries, :users, column: :user_id
