+++ db/migrate/20211005134702_create_chats.rb
@@ -1,8 +1,8 @@
 class CreateChats < ActiveRecord::Migration[5.2]
   def change
     create_table :chats do |t|
-      t.integer "user_id", null: false
-      t.integer "chat_room_id", null: false
+      t.bigint "user_id", null: false
+      t.bigint "chat_room_id", null: false
       t.text "chat", null: false
       t.timestamps
     end
